refactor: move TeamAccount to Bank folder
